chore: Move new ImageFeaturizer to onnx namespace (#1711)
