+++ deep-learning/src/main/scala/com/microsoft/azure/synapse/ml/onnx/ImageFeaturizer.scala
@@ -1,14 +1,13 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.cntk
+package com.microsoft.azure.synapse.ml.onnx
 
 import com.microsoft.azure.synapse.ml.codegen.Wrappable
 import com.microsoft.azure.synapse.ml.core.contracts.{HasInputCol, HasOutputCol}
 import com.microsoft.azure.synapse.ml.core.schema.DatasetExtensions
 import com.microsoft.azure.synapse.ml.core.schema.DatasetExtensions.findUnusedColumnName
 import com.microsoft.azure.synapse.ml.logging.BasicLogging
-import com.microsoft.azure.synapse.ml.onnx.{ONNXHub, ONNXModel, ONNXModelInfo}
 import com.microsoft.azure.synapse.ml.opencv.ImageTransformer
 import com.microsoft.azure.synapse.ml.param.TransformerParam
 import org.apache.spark.ml.linalg.DenseVector
