+++ deep-learning/src/test/scala/com/microsoft/azure/synapse/ml/onnx/ImageFeaturizerSuite.scala
@@ -1,7 +1,7 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.cntk
+package com.microsoft.azure.synapse.ml.onnx
 
 import breeze.linalg.argtopk
 import com.microsoft.azure.synapse.ml.Secrets
